kmk/read.c: more free() avoidance, this time some very small buffers for conditionals. Use static buffers for the first chunk, this avoids all alloc/free for a typical kBuild run. the cost is 16 bytes of stack per recursion.
git-svn-id: http://svn.netlabs.org/repos/kbuild/trunk@1929 8c291193-6614-0410-a08d-adc6e07170bb

+++ src/kmk/expand.c
@@ -1000,7 +1000,7 @@ allocated_variable_append (const struct variable *v)
 
   variable_buffer = 0;
 
-  assert (v->length == strlen (v->name));
+  assert ((unsigned int)v->length == strlen (v->name));
   val = variable_append (v->name, strlen (v->name), current_variable_set_list);
   variable_buffer_output (val, "", 1);
   val = variable_buffer;
